refactor move auctionResultForYouContainer to AuctionResultForYou component
